USE_O3_OPTIMIZATION Flag
This commit will give you the option to use some of my -O3 optimizations with some warning disabling flags and just a few others not included in -O3 that I think are nice.

To use place USE_O3_OPTIMIZATIONS := true or even simply USE_O3_OPTIMIZATIONS=true in your device's BoardConfig.mk.

I did not include -O3 on TARGET_thumb_CFLAGS because that is usually the reason the build will fail on some devices and really just adds lots of bloat without a noticeable difference. You can always change that flag to -O3 if you desire.

This commit also will bump your host toolchain version from 4.6 to 4.8 (which is already included in the manifest for 5.0.1) Since Host 4.8 is already here we may as well use it.

Any questions please refer them to me on hangouts.

Signed-off-by: Chet Kener <Cl3Kener@gmail.com>